Add files via upload
CellSplitter:

Splits the file into individual .fastq files for each cell based on their Barcode which they should have received from the sequencer. This assumes that the barcode is integrated into the header using UMI-tools and that it is separated by an underscore ("_"). Use it by simply activating the script and then inserting the file which are to be split.